+++ cli/src/lib.rs
@@ -1,0 +1,101 @@
+pub mod args;
+pub mod col;
+pub mod col_expr;
+pub mod cols;
+pub mod csv;
+pub mod filter;
+pub mod json;
+pub mod list_cols;
+pub mod normal;
+pub mod order;
+pub mod sorting;
+pub mod table;
+pub mod units;
+
+use {
+    crate::{
+        args::*,
+        normal::*,
+    },
+    clap::{CommandFactory, Parser},
+    std::{
+        fs,
+        os::unix::fs::MetadataExt,
+    },
+};
+
+static INTRO: &str = "
+
+**dysk** displays filesystem information in a pretty table.
+Complete documentation at https://dystroy.org/dysk
+
+";
+
+
+#[allow(clippy::match_like_matches_macro)]
+pub fn run() {
+    let args = Args::parse();
+    if args.version {
+        println!("dysk {}", env!("CARGO_PKG_VERSION"));
+        return;
+    }
+    if args.help {
+        clap_help::Printer::new(Args::command())
+            .with("introduction", INTRO)
+            .without("author")
+            .print_help();
+        return;
+    }
+    if args.list_cols {
+        list_cols::print(args.color());
+        return;
+    }
+    let mut options = lfs_core::ReadOptions::default();
+    options.remote_stats(args.remote_stats.unwrap_or_else(||true));
+    let mut mounts = match lfs_core::read_mounts(&options) {
+        Ok(mounts) => mounts,
+        Err(e) => {
+            eprintln!("Error reading mounts: {}", e);
+            return;
+        }
+    };
+    if !args.all {
+        mounts.retain(is_normal);
+    }
+    if let Some(path) = &args.path {
+        let md = match fs::metadata(path) {
+            Ok(md) => md,
+            Err(e) => {
+                eprintln!("Can't read {:?} : {}", path, e);
+                return;
+            }
+        };
+        let dev = lfs_core::DeviceId::from(md.dev());
+        mounts.retain(|m| m.info.dev == dev);
+    }
+    args.sort.sort(&mut mounts);
+    let mounts = match args.filter.clone().unwrap_or_default().filter(&mounts) {
+        Ok(mounts) => mounts,
+        Err(e) => {
+            eprintln!("Error in filter evaluation: {}", e);
+            return;
+        }
+    };
+    if args.csv {
+        csv::print(&mounts, &args).expect("writing csv failed");
+        return;
+    }
+    if args.json {
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&json::output_value(&mounts, args.units)).unwrap()
+        );
+        return;
+    }
+    if mounts.is_empty() {
+        println!("no mount to display - try\n    dysk -a");
+        return;
+    }
+    table::print(&mounts, args.color(), &args);
+}
+